Fixed _timeDelta Calculation
Fixed _timeDelta calculation bug.

+++ CodeOptPro/Assets/KamranWali/CodeOptPro/Scripts/Managers/UpdateManagerLocal.cs
@@ -68,7 +68,7 @@ namespace KamranWali.CodeOptPro.Managers
         /// <summary>
         /// This method calculates the time delta for the update manager.
         /// </summary>
-        private void CalculateTimeDelta() => _timeDelta = _objects.Count / _numUpdate;
+        private void CalculateTimeDelta() => _timeDelta = ((float) _objects.Count) / ((float) _numUpdate);
 
         /// <summary>
         /// This method calculates the actual number of objects to update.
